+++ src/RoutesHostServer/Global.asax.cs
@@ -5,6 +5,8 @@ using System.Web;
 using System.Web.Mvc;
 using System.Web.Routing;
 
+using Ariane;
+
 namespace RoutesHostServer
 {
 	public class WebApiApplication : System.Web.HttpApplication
@@ -17,6 +19,13 @@ namespace RoutesHostServer
 			{
 				Services.RoutesProvider.Current.Hydrate(repositoryFolder);
 			}
+
+			var topicName = System.Configuration.ConfigurationManager.AppSettings["TopicName"];
+
+			// Services.RouteSynchronizer.Current.Bus.Register.AddAzureTopicWriter("RoutesHostAction");
+			Services.RouteSynchronizer.Current.Bus.Register.AddAzureTopicReader("RoutesHostAction", topicName, typeof(Services.RouteMessageReader));
+
+			Services.RouteSynchronizer.Current.Bus.StartReading();
 		}
 
 		protected void Application_End()
@@ -26,6 +35,8 @@ namespace RoutesHostServer
 			{
 				Services.RoutesProvider.Current.Flush(repositoryFolder);
 			}
+
+			Services.RouteSynchronizer.Current.Bus.StopReading();
 		}
 	}
 }